# bridge/openapi/action-schema.yaml
openapi: 3.1.0
info:
  title: ChatGPT Web-first Bridge Legacy Actions
  version: 2.0.0
  description: |
    This OpenAPI file is kept only as a legacy compatibility layer.
    The primary product path is ChatGPT Custom MCP against /mcp.
servers:
  - url: https://YOUR-TUNNEL-DOMAIN.example.com
security:
  - bearerAuth: []
paths:
  /health:
    get:
      operationId: health_check
      summary: Check basic bridge health.
      security: []
      responses:
        '200':
          description: Health response
  /config:
    get:
      operationId: get_bridge_config
      summary: Read current bridge runtime and permission settings.
      responses:
        '200':
          description: Bridge config
  /projects:
    get:
      operationId: list_projects
      summary: List registered projects.
      responses:
        '200':
          description: Project list
    post:
      operationId: register_project
      summary: Register a project path.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path: { type: string }
                displayName: { type: string }
                allowShell: { type: boolean }
      responses:
        '201':
          description: Registered project
  /tasks:
    get:
      operationId: list_tasks
      summary: List tasks.
      responses:
        '200':
          description: Task list
    post:
      operationId: create_task
      summary: Create a task.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId, taskGoal]
              properties:
                projectId: { type: string }
                taskTitle: { type: string }
                taskGoal: { type: string }
      responses:
        '201':
          description: Created task
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer